[ADD] idk but lots of things

+++ src/backend/src/main/java/com/m1csc/db/backend/Repositories/CategoryRepository.java
@@ -4,10 +4,10 @@ import com.m1csc.db.backend.Entities.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigInteger;
+
 
 public interface CategoryRepository extends
-        JpaRepository<CategoryEntity, BigInteger>,
+        JpaRepository<CategoryEntity, Long>,
         JpaSpecificationExecutor<CategoryEntity> {
 
 }
