+++ src/backend/src/main/java/com/m1csc/db/backend/Repositories/EmployeeRepository.java
@@ -4,10 +4,10 @@ import com.m1csc.db.backend.Entities.EmployeeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigInteger;
+
 
 public interface EmployeeRepository extends
-        JpaRepository<EmployeeEntity, BigInteger>,
+        JpaRepository<EmployeeEntity, Long>,
         JpaSpecificationExecutor<EmployeeEntity>
 {
 }
