+++ src/backend/src/main/java/com/m1csc/db/backend/Repositories/OrderDetailRepository.java
@@ -1,13 +1,17 @@
 package com.m1csc.db.backend.Repositories;
 
 import com.m1csc.db.backend.Entities.OrderDetailEntity;
+import com.m1csc.db.backend.Entities.OrderEntity;
+import com.m1csc.db.backend.Entities.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigInteger;
+import java.util.List;
+
 
 public interface OrderDetailRepository extends
-        JpaRepository<OrderDetailEntity, BigInteger>,
+        JpaRepository<OrderDetailEntity, Long>,
         JpaSpecificationExecutor<OrderDetailEntity>
 {
+
 }
