+++ src/backend/src/main/java/com/m1csc/db/backend/Repositories/OrderRepository.java
@@ -4,10 +4,10 @@ import com.m1csc.db.backend.Entities.OrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigInteger;
+
 
 public interface OrderRepository extends
-        JpaRepository<OrderEntity, BigInteger>,
+        JpaRepository<OrderEntity, Long>,
         JpaSpecificationExecutor<OrderEntity>
 {
 }
