+++ src/backend/src/main/java/com/m1csc/db/backend/Repositories/ProductRepository.java
@@ -4,10 +4,10 @@ import com.m1csc.db.backend.Entities.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigInteger;
+
 
 public interface ProductRepository extends
-        JpaRepository<ProductEntity, BigInteger>,
+        JpaRepository<ProductEntity, Long>,
         JpaSpecificationExecutor<ProductEntity>
 
 {
