+++ src/backend/src/main/java/com/m1csc/db/backend/Repositories/ProductWarehouseRepository.java
@@ -4,10 +4,10 @@ import com.m1csc.db.backend.Entities.ProductWarehouseEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigInteger;
+
 
 public interface ProductWarehouseRepository extends
-        JpaRepository<ProductWarehouseEntity, BigInteger>,
+        JpaRepository<ProductWarehouseEntity, Long>,
         JpaSpecificationExecutor<ProductWarehouseEntity>
 {
 }
