+++ src/backend/src/main/java/com/m1csc/db/backend/Repositories/TransactionRepository.java
@@ -5,10 +5,10 @@ import com.m1csc.db.backend.Entities.TransactionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigInteger;
+
 
 public interface TransactionRepository extends
-        JpaRepository<TransactionEntity, BigInteger>,
+        JpaRepository<TransactionEntity, Long>,
         JpaSpecificationExecutor<TransactionEntity>
 {
 }
