+++ src/backend/src/main/java/com/m1csc/db/backend/Repositories/WarehouseRepository.java
@@ -4,9 +4,9 @@ import com.m1csc.db.backend.Entities.WarehouseEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-import java.math.BigInteger;
+
 
 public interface WarehouseRepository extends
-        JpaRepository<WarehouseEntity, BigInteger>,
+        JpaRepository<WarehouseEntity, Long>,
         JpaSpecificationExecutor<WarehouseEntity> {
 }
